python/requirements.txt
python-docx>=1.1
openpyxl>=3.1
python-pptx>=1.0
pypdf>=4.3
